QMM008 Fix bug

+++ nts.uk/uk.pr/pr.web/nts.uk.pr.web/src/main/webapp/view/qmm/008/h/qmm008.h.vm.ts
@@ -57,7 +57,9 @@ module nts.uk.pr.view.qmm008.h.viewmodel {
                 }
                 self.updateHistory();
             } else {
-                self.deleteHistory();
+                dialog.confirm({ messageId: "Msg_18" }).ifYes(function () {
+                    self.deleteHistory();
+                });
             }
         }
